typescript settings in WS
